refactor: split onboard provider selection
